+++ trunk/vcf/src/vcf/FoundationKit/FileStream.h
@@ -90,7 +90,7 @@ public:
 	/**
 	input
 	*/
-	virtual uint32 read( unsigned char* bytesToRead, uint32 sizeOfBytes );
+	virtual uint64 read( unsigned char* bytesToRead, uint64 sizeOfBytes );
 
 	/**
 	end of stream
@@ -134,7 +134,7 @@ public:
 	virtual uint64 getCurrentSeekPos() ;
 
 	//output
-	virtual uint32 write( const unsigned char* bytesToWrite, uint32 sizeOfBytes );
+	virtual uint64 write( const unsigned char* bytesToWrite, uint64 sizeOfBytes );
 
 	void open( const String& filename, const bool & append = false );
